Add versioned asset helper for cache busting
This commit adds a new helper function 'versioned_asset' for generating asset URLs with a versioned query string. It appends the last modified timestamp of the file to the asset URL to ensure cache busting. If the file doesn't exist, the original asset URL is returned. It's useful when you need to force browsers to download the latest version of a file after updating it. The new helper method and function have been added to 'Assets.php' and 'versioned-asset.php' respectively and the function is aliased in 'helper.php'. The path to 'versioned-asset.php' is added to the composer.json autoload section.